requirements.txt
pdfplumber
PyPDF2
ollama
streamlit
langchain
langchain_core
langchain_community
python-dotenv
langchain_ollama
faiss-cpu